requirements.txt
# Core HTTP & Async
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Image Processing
Pillow>=10.0.0

# APIs
flickrapi>=2.4.0
pyinaturalist>=0.19.0
pygbif>=0.6.0

# CLI & Progress
click>=8.1.0
rich>=13.0.0

# Logging
structlog>=23.0.0

# Environment
python-dotenv>=1.0.0

# Testing (Dev)
pytest>=7.4.0
pytest-asyncio>=0.21.0
responses>=0.23.0
pytest-cov>=4.1.0